Fixes implicit typecast from previous commit to halpower.cpp and addresses a new bug precipitated by hal-0.5.14 in which the battery plasmoid unconditionally reset screen brightness to 0
Patch by Nate Weibley

CCMAIL:nweibley@gmail.com
BUG:196232
BUG:217316

svn path=/trunk/KDE/kdebase/workspace/; revision=1062504

+++ solid/hal/halpower.cpp
@@ -387,7 +387,7 @@ float HalPower::brightness(const QString &device)
     if(reply.isValid() && reply.value().contains(device))
     {
         QDBusInterface deviceInterface("org.freedesktop.Hal", device, "org.freedesktop.Hal.Device.LaptopPanel", QDBusConnection::systemBus());
-        QDBusReply<double> brightnessReply = deviceInterface.call("GetBrightness");
+        QDBusReply<int> brightnessReply = deviceInterface.call("GetBrightness");
         if(!brightnessReply.isValid())
         {
             return 0.0;
@@ -408,7 +408,7 @@ float HalPower::brightness(const QString &device)
     {
         QDBusInterface deviceInterface("org.freedesktop.Hal", device, "org.freedesktop.Hal.Device.KeyboardBacklight", QDBusConnection::systemBus()); //TODO - I do not have a backlight enabled keyboard, so I'm guessing a bit here. Could someone please check this.
 
-        QDBusReply<double> brightnessReply = deviceInterface.call("GetBrightness");
+        QDBusReply<int> brightnessReply = deviceInterface.call("GetBrightness");
         if(!brightnessReply.isValid())
         {
             return 0.0;
